services/embeddings/requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu       # or faiss-gpu if you have GPU
requests
numpy
